# Add gaussian_blur_f32, use it in canny edge detector