… tests/data/truth/zebra_canny.png
… tests/data/truth/zebra_gaussian_10.png
… tests/data/truth/zebra_gaussian_3.png