There's definitely a lot of room for performance improvements here, but at least this is a big improvement over image::imagops::blur (the benchmarks containing "baseline" below).
```
test filter::test::bench_baseline_gaussian_stdev_1                      ... bench:   2,134,008 ns/iter (+/- 253,404)
test filter::test::bench_gaussian_f32_stdev_1                           ... bench:     454,124 ns/iter (+/- 103,688)
test filter::test::bench_baseline_gaussian_stdev_3                      ... bench:   6,050,597 ns/iter (+/- 329,343)
test filter::test::bench_gaussian_f32_stdev_3                           ... bench:     990,961 ns/iter (+/- 46,747)
test filter::test::bench_baseline_gaussian_stdev_10                     ... bench:  18,309,532 ns/iter (+/- 1,112,986)
test filter::test::bench_gaussian_f32_stdev_10                          ... bench:   2,937,146 ns/iter (+/- 147,866)
```

Canny performance:
```
Before:
test edges::test::bench_canny                                           ... bench:  13,110,707 ns/iter (+/- 3,225,960)
After:
test edges::test::bench_canny                                           ... bench:   4,475,897 ns/iter (+/- 341,225)
```